refactor: Move gradle parser into a class
Move the gradle parser into a new class. It's big, and about to get
bigger.

Also, test each of the gradle transformations individually, instead of
lumping them all into the same test.
