+++ packages/cli/src/cmds/agentInstaller/gradleParser.ts
@@ -1,0 +1,70 @@
+import moo from 'moo';
+
+export default class GradleParser {
+/**
+   * Parse the given gradle source, looking for a block identified by
+   * `blockIdentifier`. If found, return the offset in the source where
+   * the block ends. It not found, return undefined.
+   */
+   static findBlock(
+    gradleSrc: string,
+    blockIdentifier: string
+  ): number | undefined {
+    const lexer = moo.compile({
+      blockIdentifier,
+      comment: /\/\/.*?$/,
+      ident: /[\w.]+/,
+      lbrace: '{',
+      rbrace: '}',
+      punct: /[()":'/.,$*+=<>[\]~?\\&!|-]/,
+      space: { match: /\s+/, lineBreaks: true },
+      NL: { match: /\n/, lineBreaks: true },
+    });
+
+    lexer.reset(gradleSrc);
+
+    let token;
+    let startBuildscript = false;
+    let inBuildscript = false;
+    let braceIdx = 0;
+
+    // eslint-disable-next-line no-cond-assign
+    while ((token = lexer.next())) {
+      /* eslint-disable no-continue */
+      /* eslint-disable no-plusplus */
+      const t = token.type;
+      if (t === 'comment' || t === 'space') {
+        continue;
+      }
+      if (t === 'blockIdentifier') {
+        startBuildscript = true;
+        continue;
+      }
+      if (startBuildscript) {
+        // The buildscript token can appear in multiple places in a gradle file.
+        // We'll only be starting the buildscript block if it's immediately
+        // followed by a left brace.
+        if (t === 'lbrace') {
+          braceIdx = 1;
+          inBuildscript = true;
+        }
+        startBuildscript = false;
+        continue;
+      }
+
+      if (inBuildscript) {
+        if (t === 'lbrace') {
+          braceIdx++;
+        } else if (t === 'rbrace') {
+          braceIdx--;
+        }
+        if (braceIdx === 0) {
+          // We've found the right brace of the buildcript block, we're done.
+          return token.offset;
+        }
+      }
+    }
+
+    return undefined; // no buildscript block
+  }
+}
