+++ packages/cli/tests/unit/agentInstall/gradleInstaller.spec.ts
@@ -1,13 +1,12 @@
 import path from 'path';
 import { promises as fs } from 'fs';
-import { glob as globCallback } from 'glob';
-import { promisify } from 'util';
+import glob from 'glob';
 import sinon from 'sinon';
 import inquirer from 'inquirer';
 import { GradleInstaller } from '../../../src/cmds/agentInstaller/gradleInstaller';
 
-const glob = promisify(globCallback);
 const fixtureDir = path.join(__dirname, '..', 'fixtures', 'java', 'gradle');
+const dataDir = path.join(fixtureDir, 'data');
 
 describe('GradleInstaller', () => {
   afterEach(() => {
@@ -23,23 +22,23 @@ describe('GradleInstaller', () => {
   });
 
   describe('installAgent', () => {
-    it('transforms build.gradle as expected', async () => {
-      const dataDir = path.join(fixtureDir, 'data');
-      const gradle = new GradleInstaller('.');
-      const expectedExt = '.expected.gradle';
-      const files = await glob(path.join(dataDir, `*${expectedExt}`));
-      const tests = files.map((file) => path.basename(file, expectedExt));
-      const writeFile = sinon.stub(fs, 'writeFile');
+    const expectedExt = '.expected.gradle';
+    const files = glob.sync(path.join(dataDir, `*${expectedExt}`));
+    const tests = files.map((file) => path.basename(file, expectedExt));
 
-      sinon
-        .stub(inquirer, 'prompt')
-        .resolves({ addMavenCentral: 'Yes', userWillContinue: 'Continue' });
+    tests.forEach((test) => {
+      it('transforms build.gradle as expected', async () => {
+        const gradle = new GradleInstaller('.');
 
-      for (let i = 0; i < tests.length; ++i) {
-        const test = tests[i];
+        sinon
+          .stub(inquirer, 'prompt')
+          .resolves({ addMavenCentral: 'Yes', userWillContinue: 'Continue' });
+
         sinon
           .stub(gradle, 'buildFilePath')
           .value(path.join(dataDir, `${test}.actual.gradle`));
+
+        const writeFile = sinon.stub(fs, 'writeFile');
 
         const expected = (
           await fs.readFile(path.join(dataDir, `${test}${expectedExt}`))
@@ -49,7 +48,7 @@ describe('GradleInstaller', () => {
 
         const actual = writeFile.getCall(-1).args[1];
         expect(actual).toBe(expected);
-      }
+      });
     });
   });
 });
